backend-->done
front respone need to change

+++ fundooNotes/src/main/java/com/bridgelabz/fundooNotes/user/services/UserService.java
@@ -9,9 +9,9 @@ public interface UserService
 {
 	 List<User> findAll();
 	
-	 User findById(int theId);
+	 User findById(Long theId);
 	
 	 void save(UserDto userDTO);
 	
-	 void deleteById(int theId);
+	 void deleteById(Long theId);
 }
